+++ src/main/java/com/mr/moviecatalogue/inputobject/MovieIO.java
@@ -2,9 +2,12 @@ package com.mr.moviecatalogue.inputobject;
 
 import lombok.Data;
 
+/**
+ *
+ */
 @Data
 public class MovieIO {
     private String title;
     private String director;
-    private Integer rating;
+    private Float rating;
 }
